api/requirements.txt
alembic==1.11.1
fastapi==0.115.6
httpx==0.28.1
pillow==11.1.0
psycopg2==2.9.10
pytest==8.3.4
python-dotenv==1.0.0
python-multipart==0.0.20
pydantic==2.10.5
SQLAlchemy==1.4.46
torch==2.5.1
transformers[torch]==4.46.2
uvicorn==0.34.0